Have the MIDI manager sit between MIDI controllers and programs.
The MIDI manager reads MIDI controller input and sends it to the
programs that generate sound. To handle multiple programs, we have the
program manager instantiating new programs on demand and removing
inactive ones.

With this setup we assume that control changes happen on the port and
channel where note-on and note-off messages arrive as well. However, in
practice this might not be the case, for example when mixing multiple
MIDI controllers with different capabilities.