Add a proxy object for GUI updates.
We need the GUI to react to parameter changes received via MIDI. Since MIDI
is handled in a separate thread, this needs to be done in a thread-safe
way. Add a class with a global instance of it that mediates between the
MIDI and GUI threads.

At the moment it's very simplistic, avoiding any locking, except for the
implicit locking in the shared_ptr to the active program.