Add an 8-operator FM synth engine.
This is a rather flexible engine with 8 operators, each operator can
select between 5 waveforms, can be modulated by itself or any other
operator, and each operator can contribute to the output, whether it is
being used to modulate another operator or not. The envelopes work
similar to those of the DX7. Apart from that, the final output goes
through a configurable state variable filter.