Better handling of missing config file.
Ensure we create a new and empty config file if no one exists. If the
data directory is not specified, and the default data directory does not
exist, try if there is a valid data directory in the current path or any
of its parents. This allows quickly testing out Pling after compiling it
without having to install it first.